requirements.txt
streamlit==1.35.0
pandas
numpy
sqlalchemy==2.0.30
plotly==5.22.0
python-dotenv==1.0.1
jupyter
